Consolidate snippet builder logic
